feat(qr): add webrtc scanner with torch
